[AS-606] move Asset entity
