+++ src/domain/aggregates/Wallet/index.ts
@@ -1,8 +1,8 @@
 import Secp256k1KeysManagerAdapter from "../../../infrastructure/adapters/KeysManager";
 import type { IKeyManager } from "../../services/KeyManagement";
 import { deriveAddressFromPublicKey } from "../../services/AddressDerivation";
-import type Asset from "../../Asset";
-import type { Assets } from "../../Asset";
+import type Asset from "./Asset";
+import type { Assets } from "./Asset";
 import CryptoService, {
     type EncryptedData,
 } from "../../../infrastructure/adapters/Crypto";
