+++ src/domain/index.ts
@@ -1,4 +1,4 @@
-export * from "./Asset";
+export * from "./aggregates/Wallet/Asset";
 export * from "./aggregates/Wallet";
 export * from "../infrastructure/adapters/AxiosHttpClient";
 export * from "../infrastructure/adapters/BrowserStorage";
@@ -16,5 +16,5 @@ export { default as EncryptedRecord } from "./EncryptedRecord";
 export { default as SecureStorage } from "../infrastructure/adapters/BrowserStorage";
 export { default as AxiosHttpClient } from "../infrastructure/adapters/AxiosHttpClient"; 
 export { default as Wallet } from "./aggregates/Wallet";
-export { default as Asset } from "./Asset";
+export { default as Asset } from "./aggregates/Wallet/Asset";
 export { default as ReservationRecord } from "./aggregates/Reservation";
