+++ tests/security/key-management-coverage.test.ts
@@ -1,7 +1,7 @@
 import assert from "node:assert/strict";
 import test, { afterEach } from "node:test";
 
-import Asset from "../../src/domain/Asset";
+import Asset from "../../src/domain/aggregates/Wallet/Asset";
 import Wallet, { Address } from "../../src/domain/aggregates/Wallet";
 import KeyDerivationService from "../../src/services/KeyDerivation";
 import KeysManager from "../../src/services/KeysManager";
